correccion longitud codigo usuario y registro dinamico usuarios carpark

+++ database/migrations/carpark/2017_10_05_144442_create_usuarios_table.php
@@ -14,7 +14,7 @@ class CreateUsuariosTable extends Migration
     public function up()
     {
         Schema::connection('carpark')->create('TBL_Carpark_Usuarios', function (Blueprint $table) {
-            $table->integer('PK_CU_Codigo')->unsigned()->unique()->primary();
+            $table->bigInteger('PK_CU_Codigo')->unsigned()->unique()->primary();
             $table->bigInteger('CU_Cedula');
             $table->String('CU_Nombre1', 50);
             $table->String('CU_Nombre2', 50)->nullable();
